src/requirements.txt
torch
transformers
sentencepiece
ohmeow-blurr==0.0.20
datasets
bert-score
fastai